Task_4/blockchain_monitor/requirements.txt
Django
djangorestframework
django-cors-headers
psycopg2-binary
python-dotenv
celery
redis
requests
gunicorn
djangorestframework-simplejwt